+++ bundled/tool/robotcode/__main__.py
@@ -25,11 +25,6 @@ if __name__ == "__main__":
     if module_dir not in sys.path:
         sys.path.insert(0, module_dir)
 
-    if os.getenv("NO_TEAMCITY", '0').lower() not in ('true', '1', 'yes'):
-        # Use module name with dot notation
-        insert_pos = 2
-        sys.argv.insert(insert_pos, "--listener=smt_listeners.RobotIntellijListener")
-
     from robotcode.cli import robotcode
 
     robotcode(