+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/debugger/dap/RobotDebugProcolSynchronizer.java
@@ -1,0 +1,8 @@
+package dev.xeonkryptos.xeonrobotframeworkplugin.debugger.dap;
+
+import java.util.concurrent.CompletableFuture;
+
+interface RobotDebugProcolSynchronizer {
+
+    CompletableFuture<Void> robotSync();
+}