+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/debugger/dap/RobotDebugProtocolServer.java
@@ -1,0 +1,13 @@
+package dev.xeonkryptos.xeonrobotframeworkplugin.debugger.dap;
+
+import org.eclipse.lsp4j.debug.services.IDebugProtocolServer;
+import org.eclipse.lsp4j.jsonrpc.services.JsonRequest;
+
+import java.util.concurrent.CompletableFuture;
+
+public interface RobotDebugProtocolServer extends IDebugProtocolServer, RobotDebugProcolSynchronizer {
+
+    @Override
+    @JsonRequest("robot/sync")
+    CompletableFuture<Void> robotSync();
+}