+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/execution/RobotRunner.java
@@ -4,6 +4,7 @@ import com.intellij.execution.configurations.RunProfile;
 import com.intellij.execution.configurations.RunProfileState;
 import com.intellij.execution.configurations.RunnerSettings;
 import com.intellij.execution.executors.DefaultRunExecutor;
+import com.intellij.execution.process.ProcessHandler;
 import com.intellij.execution.runners.AsyncProgramRunner;
 import com.intellij.execution.runners.ExecutionEnvironment;
 import com.intellij.execution.ui.RunContentDescriptor;
@@ -43,7 +44,16 @@ public class RobotRunner extends AsyncProgramRunner<RunnerSettings> {
             if (state instanceof RobotCommandLineState) {
                 state = new RobotPythonCommandLineState(((RobotCommandLineState) state).getRobotRunConfiguration(), env);
             }
-            return super.execute(env, state);
+            RobotPythonCommandLineState robotPythonState = (RobotPythonCommandLineState) state;
+            return super.execute(env, robotPythonState).then(contentDescriptor -> {
+                if (contentDescriptor != null) {
+                    ProcessHandler processHandler = contentDescriptor.getProcessHandler();
+                    if (processHandler != null) {
+                        robotPythonState.initRobotDebugCommunicatorProcess(processHandler);
+                    }
+                }
+                return contentDescriptor;
+            });
         }
     }
 }