+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/inspections/compilation/RobotImportAnnotator.java
@@ -1,5 +1,6 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.inspections.compilation;
 
+import com.intellij.codeInspection.ProblemHighlightType;
 import com.intellij.lang.annotation.AnnotationHolder;
 import com.intellij.lang.annotation.Annotator;
 import com.intellij.lang.annotation.HighlightSeverity;
@@ -24,7 +25,10 @@ public class RobotImportAnnotator implements Annotator {
         RobotPositionalArgument positionalArgument = importElementIdentifier.getPositionalArgument();
         PsiReference reference = positionalArgument.getReference();
         if (reference.resolve() == null) {
-            holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.message("annotation.import.not-found")).range(positionalArgument).create();
+            holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.message("annotation.import.not-found"))
+                  .highlightType(ProblemHighlightType.ERROR)
+                  .range(positionalArgument)
+                  .create();
         } else {
             holder.newSilentAnnotation(HighlightSeverity.TEXT_ATTRIBUTES).textAttributes(RobotHighlighter.IMPORT_ARGUMENT).range(positionalArgument).create();
         }